Autocomplete performance improvement
Converted SQL from text to double for location, and throughout.
I probably should have done actual profiling, but it looks like a lot of
time was spent just converting strings to doubles.

So now the location is parsed immediately as a double in SQL, and is
passed that way throughout.

Also, minor cleanliness / exception checks.