Add hashcode / equals in BasicLocation for Hashmap indexing
This lets us use a BasicLocation as a key for hashmaps, and multiple
things checking to the same location will return the same thing.

So we can make caches easily.

I don't think it's actual good practice- probably you'd want the cacher
itself to figure out what things are equal.
But this is a quick cheap solution for the home-gamer.