Rudimentary Widget Support
I've added the framwork necessary to do a widget with communication with the main app. This is tricky becuase I had to do some stuff in the apple dev portal as well as synch shared app groups and what not in order to get the ios sandboxing scheme to recognize that I was allowed to share data.

Right now, clicking on a multiarrivaltrip being displayed in the bottom pane adds a thing to the today screen showing basically the name of the trip and how much time was left when you clicked it. Not tremendously helpful atm, but since I've demonstrated the technical problems are solved, I'm commiting. Next commit should hopefully make this feature usable so we can start alpha testing the ios version.

-Nighelles