Fixed Totally Whack SRDA
I don't know how to put just that change onto anything else, so Remulasce will have to change that on android if he feels like it.

Also did some stuff to fix notifications, and added notifications for when you click on a stop. I just need the nice reminder window that lets you select which arrival to add, because right now it just does the next one, which is not what its supposed to do. That should be the last fix before a usable alpha version.
Well, actually I also want to add a fix so that it selects which database to use
when you start the app, but that's potentially not required for the alpha.

-Nighelles