Added all missing features
All the features are implemented now. Nothing is pretty, but it's all working.
I also scraped all the databases. This build works for ac transit cause I need to test it. Widget is odd, I think that's because the debug toolchain isn't working, but it should work on a distribution build.

-Nighelles