Fixed Search, Regionalization
I fixed the geosearch on the IOS version, though for now the code is all in the platform specific code, so that can get fixed when I have some time to refactor. I also fixed the implementation of the LaMetroUtil to refer to a singleton RegionalizationHelper object that keeps track of the agency name. So that's working now, and I just need to do a check of which region you're in before loading the correct database.

-Nighelles