ServiceRequest makes status trip.
There's no platform UI yet- TripPopulator only kind of sort of handles
this case.

Swipe to dismiss doesn't properly scale/work. It looks like it works- but
I'm not sure if it's actually dismissing the correct items.

This won't be a problem once we get UI in for it, but TripPopulator should
really do a better job of handling Trips for which it can't make views
for.

Maybe a 'MissingView', with just a text name.