Bugfix issue 216: Bad Comparison Code
Trip priority comparator in sorting would not handle trips with equal
priority. Instead of reporting 'equal', it would select one side to be
higher.

This is incorrect, since if the priority code happened to swap the
argument order, it would receive the opposite result.

This is what happened to a user, whose instance promptly crashed.