Removed ThreadPolicy StrictMode in NotifyService #241
WillowTree guy didn't think it was necessary and wanted to know about it.
Looks like it was in the initial commit for the original, single-threaded
notify service design.

The threads now aren't doing anything that would require special
permissions. Removing the line doesn't appear to cause any problems. So,
I'm removing it.