Only hide autocomplete progress spinner if current text was filtered
The spinner used to stop too early, as the first few letters' results
would cause the spinner to halt, but the rest of the text had not been
returned.

So now wait until the actual text has been filtered.

Android doesn't actually give you enough information to do that. so we had
to override some stuff and make a listener, etc.