Allow notifications to be dismissed on 4.0
Old devices don't allow the notification action for dismiss
So we coopt the regular action and dismiss it.

The 'old device' check wasn't working right
I had thought actions were added in 4.0
But actually it was in 4.1
So we need to extend the hack to 4.0 devices.
Which I have done.