Fixed the Searchbar
Ok, we finally have an alright looking search function on the iOS version.
It has a nice dropshadow while it's displayed, though it is fixed height because I don't feel like it's worth a crapload of effort to get it to do a dynamic height searchbar which isn't really even in the iOS spec.

That's the last of the "make it look not like shit" things to do before iOS beta.

-Nighelles