Add actual .dbs, halfway through fixing norcal stuff
So the nextrip server isn't actually returning back the stopid you sent it
with. So I'm improperly assuming some things about the stopid and filling
it in in the return from the server, which fails to match what we sent and
gets rejected.

So what needs to happen is whoever knows the stop, agency, etc of the
request needs to fill it in when it gets base arrivals back from
lametroutil.