Adding BART for regional hot alpha
BART should be good for the test tomorrow. I had to do a couple of sorta quasi-not as nice as possible hacks to get it working, but the framework is nicely built so not too much of that.

Possible problems:
Code scaling to all GTFS data?
API key is hard coded, only in one place, so shouldn't be a problem.
Tiredness caused some weak code? I don't think so, sleep is for the weak.

-Nighelles