Manual Region-switching works!
Settings dialog allows you to switch agencies on and off, updating when
any checkbox is changed.

Autocomplete now loads all installed agencies at start, but only searches
through those that are active, rechecking active list every query.

So that means for non-history stops, this works perfectly.

For stops already in recent history, they work if selected. So it's an
option we have to not show recent stops if they are not an active agency.

Remaining:
- Persist "Active Agencies" between instances
- Auto-switching regions, and related items.

Optional tasks:
- Define proper names for agencies, and show them
- Check what agencies are actually installed when setting "Installed
  agencies", instead of hardcoding them in Main