Region auto-switch deal
It's in regionalizationhelper.

The agencies themselves, including locations, are still hardcoded.
So for now if the helper sees an agency with no bounds, it will simply
always add it.

But, it does check current location against agency bounds if it exists!

Note, there's no option to do this. It happens every 60 seconds.
So there still is work to do:

- Load the actual installed agencies, with all their info
- Give option for autotoggle in settings popup
- Let the user know when regions change

Note, this commit changes the Agency definition, clearing the serialized
saved autocomplete entries.