Implemented GlobalLocationProvider on iOS
So, it is possible to implement things java expects on the ios side, as long as you do all the interface things manually. It causes runtime errors if everything isn't perfect, but it works.

So GlobalLocationProvider is implemented on ios now.

Still need to add switching system with platform dep UI code.

-Nighelles